--- cmd/consumerB/main.go
@@ -1,12 +1,0 @@
-package main
-
-import (
-	"fmt"
-
-	"github.com/MszSabab/go-kafka/pkg/kafka_consumer"
-)
-
-func main() {
-	fmt.Println("okay")
-	kafka_consumer.ReadconB()
-}
